requirements.txt
# Data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.2

# Experiment tracking
mlflow==2.16.2

# API serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# Drift detection & monitoring
evidently==0.4.30

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.1
joblib==1.4.2

# Testing
pytest==8.2.0
httpx==0.27.0

# Data download
kaggle==1.6.14
